k8s: move portforward functionality to separate sub-package
The k8s package currently is a catch-all package for all kind of
k8s-related functionality. By itself it already has a prettly large list
of (transitive) dependencies, among them packages specific to the
datapath or the underlying operating system. This becomes problematic
once cross-built tools such as Cilium CLI or Hubble CLI start depending
on pkg/k8s as is the case since the portforward functionality was added
in commit ae0d2f3fb6fd ("cilium-cli: extract portforward logic into new
PortForwarder type") and they transitively start depending on the
platform-specific pacakges such as pkg/datapath.

This commit fixes that by moving the portforward functionality to
pkg/k8s/portforward which is easy because it doesn't depend on any other
code in pkg/k8s. This should help to break these transitive
dependencies.

Towards #37598
