+++ cilium-cli/k8s/dialer.go
@@ -6,13 +6,13 @@ package k8s
 import (
 	"context"
 
-	"github.com/cilium/cilium/pkg/k8s"
+	"github.com/cilium/cilium/pkg/k8s/portforward"
 )
 
 // PortForward executes in a goroutine a port forward command.
 // To stop the port-forwarding, use the context by cancelling it
-func (c *Client) PortForward(ctx context.Context, p k8s.PortForwardParameters) (*k8s.PortForwardResult, error) {
-	return k8s.NewPortForwarder(c.Clientset, c.Config).PortForward(ctx, p)
+func (c *Client) PortForward(ctx context.Context, p portforward.PortForwardParameters) (*portforward.PortForwardResult, error) {
+	return portforward.NewPortForwarder(c.Clientset, c.Config).PortForward(ctx, p)
 }
 
 // PortForwardService executes in a goroutine a port forward command towards one of the pod behind a
@@ -20,6 +20,6 @@ func (c *Client) PortForward(ctx context.Context, p k8s.PortForwardParameters) (
 // configured on the service.
 //
 // To stop the port-forwarding, use the context by cancelling it.
-func (c *Client) PortForwardService(ctx context.Context, namespace, name string, localPort, svcPort int32) (*k8s.PortForwardServiceResult, error) {
-	return k8s.NewPortForwarder(c.Clientset, c.Config).PortForwardService(ctx, namespace, name, localPort, svcPort)
+func (c *Client) PortForwardService(ctx context.Context, namespace, name string, localPort, svcPort int32) (*portforward.PortForwardServiceResult, error) {
+	return portforward.NewPortForwarder(c.Clientset, c.Config).PortForwardService(ctx, namespace, name, localPort, svcPort)
 }
