+++ pkg/k8s/portforward/portforward.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package k8s
+package portforward
 
 import (
 	"context"
